VITKTM/vitkt_m/requirements.txt
absl-py==0.13.0
addict==2.4.0
albumentations==1.1.0
attributee==0.1.7
attrs==21.2.0
bidict==0.21.2
cachetools==4.2.2
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.3
dataclasses==0.8
decorator==4.4.2
dominate==2.6.0
easydict==1.9
future==0.18.2
google-auth==1.35.0
google-auth-oauthlib==0.4.5
grpcio==1.39.0
icc-rt==2020.0.133
idna==3.2
imagecorruptions==1.1.2
imageio==2.9.0
importlib-metadata==4.6.1
intel-numpy==1.15.1
intel-openmp==2020.0.133
joblib==1.0.1
jpeg4py==0.1.4
jsonpatch==1.32
jsonpointer==2.1
jsonschema==3.2.0
kiwisolver==1.3.1
llvmlite==0.36.0
lmdb==1.2.1
Markdown==3.3.4
matplotlib==3.3.4
mkl==2019.0
mkl-fft==1.0.6
mkl-random==1.0.1.1
networkx==2.5.1
ninja==1.10.2.3
numba==0.53.1
numpy==1.19.5
oauthlib==3.1.1
olefile==0.44
onnx==1.11.0
onnxruntime-gpu==1.6.0
opencv-python==4.5.3.56
opencv-python-headless==4.5.4.58
ordered-set==4.0.2
packaging==21.0
pandas==1.1.5
phx-class-registry==3.0.5
Pillow==8.3.1
pip==21.1.3
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.7
pycparser==2.20
PyLaTeX==1.4.1
pyparsing==2.4.7
pyrsistent==0.18.0
pytest-runner==5.3.1
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.1.0
qudida==0.0.4
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.2.1
setuptools==58.0.4
Shapely==1.7.1
six==1.16.0
sklearn==0.0
tb-nightly==2.6.0a20210728
tbb==2021.4.0
tbb4py==2021.4.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
terminaltables==3.1.0
thop==0.0.31.post2005241907
threadpoolctl==2.2.0
tifffile==2020.9.3
tikzplotlib==0.9.9
timm==0.3.2
torchfile==0.1.0
torchvision==0.8.1
tornado==6.1
tqdm==4.61.2
typing-extensions==3.10.0.0
urllib3==1.26.6
visdom==0.1.8.9
vot-toolkit==0.5.3
vot-trax==3.0.3
websocket-client==1.1.0
Werkzeug==2.0.1
wheel==0.36.2
xlwt==1.3.0
yacs==0.1.8
yapf==0.31.0
zipp==3.5.0
